+++ src/main/java/com/api/hotifi/payment/services/interfaces/ISellerPaymentService.java
@@ -3,13 +3,15 @@ package com.api.hotifi.payment.services.interfaces;
 import com.api.hotifi.identity.entities.User;
 import com.api.hotifi.payment.web.responses.SellerReceiptResponse;
 
+import java.math.BigDecimal;
+
 public interface ISellerPaymentService {
 
     //DO NOT ADD TO CONTROLLER
-    void addSellerPayment(User seller, double amountEarned);
+    void addSellerPayment(User seller, BigDecimal amountEarned);
 
     //DO NOT ADD TO CONTROLLER
-    void updateSellerPayment(User seller, double amountEarned, boolean isUpdateTimeOnly);
+    void updateSellerPayment(User seller, BigDecimal amountEarned, boolean isUpdateTimeOnly);
 
     SellerReceiptResponse withdrawSellerPayment(Long sellerId);
