+++ src/main/java/com/api/hotifi/payment/services/interfaces/ISellerReceiptService.java
@@ -4,12 +4,13 @@ import com.api.hotifi.identity.entities.User;
 import com.api.hotifi.payment.entities.SellerPayment;
 import com.api.hotifi.payment.web.responses.SellerReceiptResponse;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface ISellerReceiptService {
 
     //DO NOT ADD TO CONTROLLER
-    SellerReceiptResponse addSellerReceipt(User seller, SellerPayment sellerPayment, double sellerAmountPaid);
+    SellerReceiptResponse addSellerReceipt(User seller, SellerPayment sellerPayment, BigDecimal sellerAmountPaid);
 
     SellerReceiptResponse getSellerReceipt(Long id);
