+++ src/main/java/com/api/hotifi/payment/web/responses/BuyerStatsResponse.java
@@ -3,11 +3,13 @@ package com.api.hotifi.payment.web.responses;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class BuyerStatsResponse {
 
-    public double totalPendingRefunds;
+    public BigDecimal totalPendingRefunds;
 
     public double totalDataBought;
